make abstract level of singleCNN and ResNet to inherit the different frameworks
move original pytorch model to model collection
